manaxc-project/code/importers/requirements.txt
openpyxl==3.1.2
python-dotenv==1.0.0
supabase==2.3.0
nameparser==1.1.3
